[MLDB-1468] removed credential daemon
