+++ container_files/mldb_nginx_site.conf
@@ -10,7 +10,7 @@ upstream mldb_logger {
 
 server {
     listen 80;
-    server_name mldb; 
+    server_name mldb;
     keepalive_requests 1000000000;
 
     include /etc/nginx/uwsgi_params;
@@ -97,14 +97,6 @@ server {
             proxy_set_header Connection "";
             add_header Access-Control-Allow-Origin *;
        }
-    
-        # Credentialsd routes
-        location {{HTTP_BASE_URL}}/{{CREDENTIALSD_BASE_URL}}/types {
-            proxy_pass http://{{CREDENTIALSD_LISTEN_ADDR}}:{{CREDENTIALSD_LISTEN_PORT}}/v1/types;
-        }
-        location {{HTTP_BASE_URL}}/{{CREDENTIALSD_BASE_URL}}/rules {
-            proxy_pass http://{{CREDENTIALSD_LISTEN_ADDR}}:{{CREDENTIALSD_LISTEN_PORT}}/v1/rules;
-        }
 
         # MLDB routes
         location {{HTTP_BASE_URL}}/v1 {
@@ -138,15 +130,6 @@ server {
         }
     }
 
-    # MLDB routes at root - pymldb uses it from within container
-    # Credentialsd routes
-    location {{CREDENTIALSD_BASE_URL}}/types {
-        proxy_pass http://{{CREDENTIALSD_LISTEN_ADDR}}:{{CREDENTIALSD_LISTEN_PORT}}/v1/types;
-    }
-    location {{CREDENTIALSD_BASE_URL}}/rules {
-        proxy_pass http://{{CREDENTIALSD_LISTEN_ADDR}}:{{CREDENTIALSD_LISTEN_PORT}}/v1/rules;
-    }
- 
     # MLDB routes
     location /v1 {
         proxy_pass http://mldb_upstream/v1;
