+++ server/mldb_server.h
@@ -28,12 +28,14 @@ struct PluginCollection;
 struct DatasetCollection;
 struct ProcedureCollection;
 struct FunctionCollection;
+struct CredentialRuleCollection;
 struct TypeClassCollection;
 
 struct Plugin;
 struct Dataset;
 struct Procedure;
 struct Function;
+struct CredentialRule;
 
 struct MatrixNamedRow;
 
@@ -47,7 +49,7 @@ struct MatrixNamedRow;
 */
 
 struct MldbServer: public ServicePeer, public EventRecorder {
-    
+
     MldbServer(const std::string & serviceName = "mldb",
                const std::string & etcdUri = "",
                const std::string & etcdPath = "",
@@ -88,7 +90,7 @@ struct MldbServer: public ServicePeer, public EventRecorder {
               std::string staticFilesPath = "file://mldb/container_files/public_html/resources",
               std::string staticDocPath = "file://mldb/container_files/public_html/doc",
               bool hideInternalEntities = false);
-    
+
     void start();
 
     void shutdown();
@@ -96,13 +98,14 @@ struct MldbServer: public ServicePeer, public EventRecorder {
     typedef std::function<bool (const Json::Value & progress)> OnProgress;
 
     /** Obtain the dataset with the given configuration. */
-    
+
     std::shared_ptr<RestRouteManager> routeManager;
 
     std::shared_ptr<PluginCollection> plugins;
     std::shared_ptr<DatasetCollection> datasets;
     std::shared_ptr<ProcedureCollection> procedures;
     std::shared_ptr<FunctionCollection> functions;
+    std::shared_ptr<CredentialRuleCollection> credentials;
     std::shared_ptr<TypeClassCollection> types;
 
     /** Parse and perform an SQL query. */
@@ -143,7 +146,7 @@ struct MldbServer: public ServicePeer, public EventRecorder {
 private:
     void preInit();
     bool initRoutes();
-    void initCollections(std::string configurationPath,
+    void initCollections(std::string credentialsPath,
                          std::string staticFilesPath,
                          std::string staticDocPath,
                          bool hideInternalEntities);
